feat: update packages and unlock cargo hash
