--- config/src/sandbox/linux/mod.rs
@@ -1,1 +1,0 @@
-pub mod paths;
