--- config/src/sandbox/linux/paths.rs
@@ -1,48 +1,0 @@
-// use vorpal_schema::vorpal::package::v0::PackageSandboxPath;
-
-// pub fn add_glibc() -> Vec<PackageSandboxPath> {
-//     vec![
-//         PackageSandboxPath {
-//             source: "/lib/aarch64-linux-gnu/ld-linux-aarch64.so.1".to_string(),
-//             target: "/lib/ld-linux-aarch64.so.1".to_string(),
-//         },
-//         PackageSandboxPath {
-//             source: "/lib/aarch64-linux-gnu/libc.so.6".to_string(),
-//             target: "/lib/aarch64-linux-gnu/libc.so.6".to_string(),
-//         },
-//     ]
-// }
-
-// pub fn add_bash() -> Vec<PackageSandboxPath> {
-//     vec![
-//         PackageSandboxPath {
-//             source: "/bin/bash".to_string(),
-//             target: "/bin/bash".to_string(),
-//         },
-//         PackageSandboxPath {
-//             source: "/bin/bashbug".to_string(),
-//             target: "/bin/bashbug".to_string(),
-//         },
-//     ]
-// }
-
-// pub fn add_ncurses() -> Vec<PackageSandboxPath> {
-//     vec![PackageSandboxPath {
-//         source: "/lib/aarch64-linux-gnu/libtinfo.so.6".to_string(),
-//         target: "/lib/aarch64-linux-gnu/libtinfo.so.6".to_string(),
-//     }]
-// }
-
-// pub fn add_resolv_conf() -> Vec<PackageSandboxPath> {
-//     vec![PackageSandboxPath {
-//         source: "/etc/resolv.conf".to_string(),
-//         target: "/etc/resolv.conf".to_string(),
-//     }]
-// }
-
-// pub fn add_ssl_certs() -> Vec<PackageSandboxPath> {
-//     vec![PackageSandboxPath {
-//         source: "/etc/ssl/certs".to_string(),
-//         target: "/etc/ssl/certs".to_string(),
-//     }]
-// }
