--- config/src/sandbox/mod.rs
@@ -1,2 +1,0 @@
-pub mod linux;
-pub mod scripts;
