+++ sdk/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "vorpal-sdk"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = { default-features = false, version = "1" }
+clap = { default-features = false, features = ["color", "derive", "error-context", "help", "std", "suggestions", "usage"], version = "4" }
+indoc = { default-features = false, version = "2" }
+serde = { features = ["serde_derive"], version = "1" }
+serde_json = { default-features = false, features = ["std"], version = "1" }
+sha256 = { default-features = false, version = "1" }
+tonic = { default-features = false, version = "0" }
+tracing = { default-features = false, version = "0" }
+vorpal-schema = { default-features = false, path = "../schema" }
+vorpal-store = { default-features = false, path = "../store" }
