+++ sdk/src/config/artifact/toolchain/cargo.rs
@@ -1,5 +1,5 @@
-use crate::{
-    artifact::{build_artifact, step_env_artifact},
+use crate::config::{
+    artifact::{add_artifact, get_artifact_envkey},
     ContextConfig,
 };
 use anyhow::{bail, Result};
@@ -19,7 +19,7 @@ pub fn artifact(context: &mut ContextConfig) -> Result<ArtifactId> {
         X8664Macos.into(),
     ];
 
-    let source = build_artifact(
+    let source = add_artifact(
         context,
         vec![],
         vec![],
@@ -42,14 +42,14 @@ pub fn artifact(context: &mut ContextConfig) -> Result<ArtifactId> {
         systems.clone(),
     )?;
 
-    build_artifact(
+    add_artifact(
         context,
         vec![source.clone()],
         vec![],
         name,
         format!(
             "cp -prv {}/cargo/. \"$VORPAL_OUTPUT\"/",
-            step_env_artifact(&source)
+            get_artifact_envkey(&source)
         ),
         vec![],
         systems,
