+++ sdk/src/config/artifact/toolchain/protoc.rs
@@ -1,5 +1,5 @@
-use crate::{
-    artifact::{build_artifact, step_env_artifact},
+use crate::config::{
+    artifact::{add_artifact, get_artifact_envkey},
     ContextConfig,
 };
 use anyhow::{bail, Result};
@@ -19,7 +19,7 @@ pub fn artifact(context: &mut ContextConfig) -> Result<ArtifactId> {
         X8664Macos.into(),
     ];
 
-    let source = build_artifact(
+    let source = add_artifact(
         context,
         vec![],
         vec![],
@@ -42,7 +42,7 @@ pub fn artifact(context: &mut ContextConfig) -> Result<ArtifactId> {
         systems.clone(),
     )?;
 
-    build_artifact(
+    add_artifact(
         context,
         vec![source.clone()],
         vec![],
@@ -52,7 +52,7 @@ pub fn artifact(context: &mut ContextConfig) -> Result<ArtifactId> {
             cp -r {source}/include \"$VORPAL_OUTPUT/include\"
 
             chmod +x \"$VORPAL_OUTPUT/bin/protoc\"",
-            source = step_env_artifact(&source),
+            source = get_artifact_envkey(&source),
         },
         vec![],
         systems,
